+++ src/pipeline_expansion/algorithm.py
@@ -15,7 +15,7 @@ from data_exporter.dap_to_expansion import (
 )
 from data_model import NodeEdgeModel, EdgeData
 from pipeline_expansion.admm_helpers import ADMM, ADMMResult
-from pipeline_expansion.api import run_sddp, generate_scenarios
+from api.sddp import run_sddp, generate_scenarios
 from data_model.sddp import (
     AdditionalParams,
     BenderCut,
